requirements.txt
functions-framework==3.0.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.63.0
pytz==2022.1
